# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/manage/cats": {
            "post": {
                "operationId": "CatController_createCat",
                "summary": "创建猫咪",
                "description": "创建猫咪,需要管理员权限",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCatBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateCatResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/manage/cat/{id}": {
            "put": {
                "operationId": "CatController_updateCat",
                "summary": "修改猫咪信息",
                "description": "修改猫咪信息,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "猫咪ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCatBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCatResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "CommentController_deleteComment",
                "summary": "删除评论",
                "description": "删除评论,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "评论ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCommentResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/manage/cats/photos": {
            "get": {
                "operationId": "CatPhotoController_getPhotos",
                "summary": "获取所有猫咪照片",
                "description": "获取所有猫咪照片,需要管理员权限",
                "parameters": [
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,限制数量",
                        "schema": {
                            "minimum": 1,
                            "maximum": 30,
                            "type": "number"
                        }
                    },
                    {
                        "name": "start",
                        "required": true,
                        "in": "query",
                        "description": "分页参数,开始位置",
                        "schema": {
                            "minimum": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GetCatsPhotosResponseDto"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/manage/cats/photo/{id}": {
            "put": {
                "operationId": "CatPhotoController_updatePhoto",
                "summary": "修改猫咪照片信息",
                "description": "修改猫咪照片信息,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "照片ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCatPhotoBodyDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateCatPhotoResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "CatPhotoController_deletePhoto",
                "summary": "删除猫咪照片",
                "description": "删除猫咪照片,需要管理员权限",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "照片ID",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCatPhotoResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "tags": [
                    "管理"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Kitty后端APi服务",
        "description": {
            "$ref": "./index.md"
        },
        "version": "1.15",
        "contact": {},
        "x-logo": {
            "url": "./logo.png",
            "backgroundColor": "#FFFFFF"
        }
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "统一认证",
                "type": "http"
            }
        },
        "schemas": {
...
            "DeleteCommentResponseDto": {
                "type": "object",
                "properties": {}
            },
...
            "CreateCatBodyDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "猫咪名称"
                    },
                    "species": {
                        "type": "string",
                        "description": "猫咪物种"
                    },
                    "isNeuter": {
                        "type": "boolean",
                        "description": "猫咪是否已经绝育"
                    },
                    "description": {
                        "type": "string",
                        "description": "猫咪描述"
                    },
                    "haunt": {
                        "type": "string",
                        "description": "猫咪出没地点"
                    },
                    "status": {
                        "type": "string",
                        "description": "猫咪状态,0表示失踪,1表示正常,2表示生病,3表示住院,4表示被领养,5表示死亡",
                        "enum": [
                            "0",
                            "1",
                            "2",
                            "3",
                            "4",
                            "5"
                        ]
                    }
                },
                "required": [
                    "name",
                    "species",
                    "isNeuter",
                    "description",
                    "haunt",
                    "status"
                ]
            },
            "CreateCatResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "猫咪ID",
                        "minimum": 1
                    }
                },
                "required": [
                    "id"
                ]
            },
            "UpdateCatBodyDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "猫咪名称"
                    },
                    "species": {
                        "type": "string",
                        "description": "猫咪物种"
                    },
                    "isNeuter": {
                        "type": "boolean",
                        "description": "猫咪是否已经绝育"
                    },
                    "description": {
                        "type": "string",
                        "description": "猫咪描述"
                    },
                    "haunt": {
                        "type": "string",
                        "description": "猫咪出没地点"
                    },
                    "status": {
                        "type": "string",
                        "description": "猫咪状态,0表示失踪,1表示正常,2表示生病,3表示住院,4表示被领养,5表示死亡",
                        "enum": [
                            "0",
                            "1",
                            "2",
                            "3",
                            "4",
                            "5"
                        ]
                    }
                },
                "required": [
                    "name",
                    "species",
                    "isNeuter",
                    "description",
                    "haunt",
                    "status"
                ]
            },
            "UpdateCatResponseDto": {
                "type": "object",
                "properties": {}
            },
...
            "GetCatsPhotosResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "description": "照片ID",
                        "minimum": 1
                    },
                    "type": {
                        "type": "string",
                        "description": "照片状态,0表示封面图,1表示精选图,2表示其他,3表示正在审核",
                        "enum": [
                            "0",
                            "1",
                            "2",
                            "3"
                        ]
                    },
                    "rawFileName": {
                        "type": "string",
                        "description": "原图图片文件名"
                    },
                    "fileName": {
                        "type": "string",
                        "description": "展示图片文件名",
                        "nullable": true
                    },
                    "createdDate": {
                        "format": "date-time",
                        "type": "string",
                        "description": "图片发布时间"
                    },
                    "comment": {
                        "type": "string",
                        "description": "照片评论"
                    },
                    "userID": {
                        "type": "number",
                        "description": "照片发布者ID"
                    }
                },
                "required": [
                    "id",
                    "type",
                    "rawFileName",
                    "fileName",
                    "createdDate",
                    "comment",
                    "userID"
                ]
            },
            "UpdateCatPhotoBodyDto": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "description": "照片状态,0表示封面图,1表示精选图,2表示其他,3表示正在审核",
                        "enum": [
                            "0",
                            "1",
                            "2",
                            "3"
                        ]
                    }
                },
                "required": [
                    "type"
                ]
            },
            "UpdateCatPhotoResponseDto": {
                "type": "object",
                "properties": {}
            },
            "DeleteCatPhotoResponseDto": {
                "type": "object",
                "properties": {}
            },
...
            "HttpExceptionResponseDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number",
                        "description": "HTTP状态码"
                    },
                    "msg": {
                        "type": "object",
                        "description": "错误信息"
                    }
                },
                "required": [
                    "status",
                    "msg"
                ]
            }
        },
        "responses": {
            "400": {
                "description": "参数错误",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            },
            "401": {
                "description": "Authorization字段无效",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            },
            "403": {
                "description": "无权限",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/HttpExceptionResponseDto"
                        }
                    }
                }
            }
        }
    }
}